Level 3 helper script
This generates a file with contents to be used as the string argument to the
level03 program. With the index argument set to -28, the function array ends up
referencing the beginning of our string, which just to happens to point at the
deprecated `run` function. The first command that gets run on the system is
nonsense, but then we print the contents of /home/level04/.password. We could
just have easily invoked bash and had a full console as level04.